rename my_user.py to user.py
